// app/lib/definitions/enums/deities/PatronoMetaField.ts
enum PatronoMetaField {
  nome = "nome",
  titoloPatrono = "titolopatrono",
  tipoPatrono = "tipopatrono",
  gradoPatrono = "gradopatrono",
  card = "card",
  astri = "astri",
  elemento = "elemento",
  classe = "classe",
  festivita = "festivita",
  colore = "colore",
  tradizione = "tradizione",
  allineamento = "allineamento",
  dominioAllineamento = "dominioallineamento",
  residenza = "residenza",
  luogo = "luogo",
  significato = "significato",
}

export default PatronoMetaField;